@@ -7,58 +7,58 @@
   "main": "gatsby-browser.js",
   "dependencies": {
     "@popperjs/core": "^2.11.5",
+    "@sindresorhus/slugify": "^2.1.0",
     "bootstrap": "^5.1.3",
-    "bootstrap-icons": "^1.8.2",
-    "clsx": "^1.1.1",
-    "gatsby": "^1.0.1",
-    "gatsby-plugin-feed": "^2.13.1",
-    "gatsby-plugin-google-analytics": "^2.11.0",
-    "gatsby-plugin-graphql-codegen": "^3.1.0",
-    "gatsby-plugin-image": "^0.7.2",
-    "gatsby-plugin-layout": "^1.10.0",
-    "gatsby-plugin-manifest": "^2.6.1",
-    "gatsby-plugin-offline": "^3.10.2",
+    "bootstrap-icons": "^1.8.3",
+    "clsx": "^1.2.1",
+    "gatsby": "^4.18.0",
+    "gatsby-plugin-feed": "^4.18.0",
+    "gatsby-plugin-google-analytics": "^4.18.0",
+    "gatsby-plugin-image": "^2.18.0",
+    "gatsby-plugin-layout": "^3.18.0",
+    "gatsby-plugin-manifest": "^4.18.0",
+    "gatsby-plugin-offline": "^5.18.0",
     "gatsby-plugin-purgecss": "^6.1.2",
-    "gatsby-plugin-react-helmet": "^3.10.0",
-    "gatsby-plugin-sass": "^3.2.0",
+    "gatsby-plugin-react-helmet": "^5.18.0",
+    "gatsby-plugin-sass": "^5.18.0",
     "gatsby-plugin-scroll-reveal": "0.0.7",
-    "gatsby-plugin-sharp": "^2.14.4",
-    "gatsby-plugin-sitemap": "^2.12.0",
-    "gatsby-remark-autolink-headers": "^2.11.0",
-    "gatsby-remark-images": "^3.11.1",
-    "gatsby-remark-katex": "^3.10.0",
-    "gatsby-remark-prismjs": "^3.13.0",
-    "gatsby-remark-smartypants": "^2.10.0",
-    "gatsby-source-filesystem": "^2.11.1",
-    "gatsby-transformer-remark": "^5.14.0",
-    "gatsby-transformer-sharp": "^2.12.1",
-    "gatsby-transformer-yaml": "^1.0.1",
-    "katex": "^0.15.6",
+    "gatsby-plugin-sharp": "^4.18.0",
+    "gatsby-plugin-sitemap": "^5.18.0",
+    "gatsby-remark-autolink-headers": "^5.18.0",
+    "gatsby-remark-images": "^6.18.0",
+    "gatsby-remark-katex": "^6.18.0",
+    "gatsby-remark-prismjs": "^6.18.0",
+    "gatsby-remark-smartypants": "^5.18.0",
+    "gatsby-source-filesystem": "^4.18.0",
+    "gatsby-transformer-remark": "^5.18.0",
+    "gatsby-transformer-sharp": "^4.18.0",
+    "gatsby-transformer-yaml": "^4.18.0",
+    "katex": "^0.16.0",
     "prismjs": "^1.28.0",
-    "react": "^18.1.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^18.1.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "sass": "1.52.1"
+    "sass": "1.53.0"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^2.6.2",
-    "@graphql-codegen/introspection": "^2.1.1",
-    "@graphql-codegen/typescript": "^2.4.11",
-    "@graphql-codegen/typescript-operations": "^2.4.0",
-    "@types/node": "^17.0.35",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.5",
+    "@graphql-codegen/cli": "^2.8.0",
+    "@graphql-codegen/introspection": "^2.2.0",
+    "@graphql-codegen/typescript": "^2.7.1",
+    "@graphql-codegen/typescript-operations": "^2.5.1",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/react-helmet": "^6.1.5",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "babel-preset-gatsby": "^2.14.0",
-    "eslint": "^8.16.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "babel-preset-gatsby": "^2.18.0",
+    "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "gh-pages": "^4.0.0",
-    "prettier": "2.6.2",
-    "typescript": "^4.6.4"
+    "prettier": "2.7.1",
+    "typescript": "^4.7.4"
   },
   "keywords": [
     "gatsby",
